+++ tests/testthat/test-utils.R
@@ -81,10 +81,78 @@ test_that("parse_artifact_name works", {
   expect_identical(unlist(res), unlist(expected))
 })
 
+test_that("create_full_artifact_name works", {
+  names <- list(
+    list("workspace", "artifact-name", "versionOrAlias"),
+    list("workspace", "artifact-name", NULL),
+    list(NULL, "artifact-name", "versionOrAlias"),
+    list(NULL, "artifact-name", NULL)
+  )
+  expected <- list(
+    "workspace/artifact-name:versionOrAlias",
+    "workspace/artifact-name",
+    "artifact-name:versionOrAlias",
+    "artifact-name"
+  )
+  res <- sapply(names, function(n) create_full_artifact_name(
+    artifact_name = n[[2]], workspace = n[[1]], version = n[[3]]
+  ))
+  expect_identical(unlist(res), unlist(expected))
+})
+
 test_that("encode_metadata and decode_metadata works", {
   metadata <- list(foo="bar")
 
   encoded <- encode_metadata(metadata)
   decoded <- decode_metadata(encoded)
   expect_equal(decoded, metadata)
+})
+
+test_that("validate_artifact_overwrite_strategy works", {
+  strategies <- list(TRUE, FALSE, "FAIL", "PRESERVE", "OVERWRITE")
+  expected <- list("OVERWRITE", "FAIL", "FAIL", "PRESERVE", "OVERWRITE")
+
+  res <- sapply(strategies, validate_artifact_overwrite_strategy)
+  expect_identical(unlist(res), unlist(expected))
+})
+
+test_that("validate_artifact_overwrite_strategy fails appropriately", {
+  expect_error(
+    validate_artifact_overwrite_strategy("NOT EXISTING"),
+    "Unsupported overwrite_strategy value: NOT EXISTING"
+  )
+  expect_error(
+    validate_artifact_overwrite_strategy(1),
+    "Unsupported overwrite_strategy value: 1"
+  )
+  expect_error(
+    validate_artifact_overwrite_strategy(NULL),
+    "Unsupported overwrite_strategy value: "
+  )
+})
+
+test_that("resolve_artifact_asset_path works for not existing file", {
+  parent_dir <- withr::local_tempdir()
+  asset_file <- file.path("asset_dir", "asset.dat")
+  result <- resolve_artifact_asset_path(parent_dir = parent_dir,
+                                        asset_file = asset_file,
+                                        overwrite_strategy = "FAIL")
+  expect_false(result$already_exists)
+  expect_equal(result$asset_path, file.path(parent_dir, asset_file))
+  expect_true(dir.exists(file.path(parent_dir, "asset_dir")))
+})
+
+test_that("resolve_artifact_asset_path works for existing file with OVERWRITE", {
+  parent_dir <- withr::local_tempdir()
+  asset_file <- "asset.dat"
+  asset_file_path <- file.path(parent_dir, asset_file)
+  cat("Test data", file = asset_file_path)
+  expect_true(file.exists(asset_file_path))
+
+  result <- resolve_artifact_asset_path(parent_dir = parent_dir,
+                                        asset_file = asset_file,
+                                        overwrite_strategy = "OVERWRITE")
+  expect_true(result$already_exists)
+  expect_equal(result$asset_path, file.path(parent_dir, asset_file))
+  expect_false(file.exists(asset_file_path))
 })